requirements.txt
requests
beautifulsoup4
google-generativeai
flask
python-dateutil